Fix imports after move of DNF payload
