+++ pyanaconda/modules/payload/factory.py
@@ -19,7 +19,7 @@
 from abc import ABC, abstractclassmethod
 
 from pyanaconda.modules.payload.base.constants import HandlerType, SourceType
-from pyanaconda.modules.payload.dnf.dnf import DNFHandlerModule
+from pyanaconda.modules.payload.payloads.dnf.dnf import DNFHandlerModule
 from pyanaconda.modules.payload.live.live_image import LiveImageHandlerModule
 from pyanaconda.modules.payload.live.live_os import LiveOSHandlerModule
 from pyanaconda.modules.payload.sources.live_os import LiveOSSourceModule
