+++ tests/nosetests/pyanaconda_tests/module_payload_dnf_test.py
@@ -25,10 +25,11 @@ from tests.nosetests.pyanaconda_tests import patch_dbus_publish_object
 
 from pyanaconda.modules.common.constants.objects import DNF_PACKAGES
 from pyanaconda.modules.common.errors import InvalidValueError
-from pyanaconda.modules.payload.dnf.packages.packages import PackagesHandlerModule
-from pyanaconda.modules.payload.dnf.packages.packages_interface import PackagesHandlerInterface
-from pyanaconda.modules.payload.dnf.packages.constants import TIMEOUT_UNSET, RETRIES_UNSET, \
-    LANGUAGES_DEFAULT, LANGUAGES_NONE
+from pyanaconda.modules.payload.payloads.dnf.packages.packages import PackagesHandlerModule
+from pyanaconda.modules.payload.payloads.dnf.packages.packages_interface import \
+    PackagesHandlerInterface
+from pyanaconda.modules.payload.payloads.dnf.packages.constants import TIMEOUT_UNSET, \
+    RETRIES_UNSET, LANGUAGES_DEFAULT, LANGUAGES_NONE
 from tests.nosetests.pyanaconda_tests.module_payload_shared import PayloadHandlerMixin
